Replaced group_by with lexsort in index creation, renamed array.py
