+++ astropy/table/sorted_array.py
